refactor: step 분리
